requirements.txt
appnope==0.1.2
asgiref==3.3.4
backcall==0.2.0
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
decorator==5.0.7
Django==3.2
google-api-core==1.26.3
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
mecab-python3==1.0.3
packaging==20.9
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
protobuf==3.15.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.8.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.4
wcwidth==0.2.5